DON-998: Refactor, introduce security service inspired somewhat by Symfony's security service https://symfony.com/doc/current/security.html#fetching-the-user-from-a-service
