+++ integrationTests/ListRegularGivingMandatesTest.php
@@ -7,10 +7,15 @@ use GuzzleHttp\Psr7\ServerRequest;
 use MatchBot\Application\Actions\RegularGivingMandate\GetAllForUser;
 use MatchBot\Application\Auth\PersonManagementAuthMiddleware;
 use MatchBot\Domain\DayOfMonth;
+use MatchBot\Domain\DonorAccount;
+use MatchBot\Domain\DonorAccountRepository;
+use MatchBot\Domain\DonorName;
+use MatchBot\Domain\EmailAddress;
 use MatchBot\Domain\Money;
 use MatchBot\Domain\PersonId;
 use MatchBot\Domain\RegularGivingMandate;
 use MatchBot\Domain\Salesforce18Id;
+use MatchBot\Domain\StripeCustomerId;
 use Ramsey\Uuid\Uuid;
 use Ramsey\Uuid\UuidInterface;
 
@@ -26,6 +31,14 @@ class ListRegularGivingMandatesTest extends IntegrationTest
     {
         parent::setUp();
         $this->donorId = PersonId::of(Uuid::uuid4()->toString());
+        $donorAccountRepo = $this->getService(DonorAccountRepository::class);
+
+        $donorAccountRepo->save(new DonorAccount(
+            $this->donorId,
+            EmailAddress::of('test@example.com'),
+            DonorName::of('first', 'last'),
+            StripeCustomerId::of('cus_' . $this->randomString())
+        ));
     }
 
     /**
