+++ src/Application/Actions/RegularGivingMandate/CreateCustomerSession.php
@@ -4,15 +4,11 @@ namespace MatchBot\Application\Actions\RegularGivingMandate;
 
 use Laminas\Diactoros\Response\JsonResponse;
 use MatchBot\Application\Actions\Action;
-use MatchBot\Application\Auth\PersonWithPasswordAuthMiddleware;
-use MatchBot\Application\Security\SecurityService;
+use MatchBot\Application\Security\Security;
 use MatchBot\Client\Stripe;
-use MatchBot\Domain\DonorAccountRepository;
-use MatchBot\Domain\PersonId;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Log\LoggerInterface;
-use Security;
 
 /**
  * Creates a stripe customer session for use with regular giving
@@ -22,7 +18,7 @@ class CreateCustomerSession extends Action
     public function __construct(
         LoggerInterface $logger,
         private Stripe $stripeClient,
-        private SecurityService $security,
+        private Security $security,
     ) {
         parent::__construct($logger);
     }
